refactor(house-number-importer): use the shared DOM builders
dot() moves into tab.ts, its only caller: it reads the feature's own STATUS_ICONS,
so it was the one builder with no business in a shared module, and ui/format.ts
states in its header that it holds no DOM. Everything else now comes from
src/ui/dom.ts, and the feature's own dom.ts is gone.

The plan gains two corrections found while running it: where dot() belongs, and a
warning never to run Prettier here — the repo has no config and 81 files under src
do not match its defaults, so formatting a touched file rewrites unrelated lines.
