+++ src/house-number-importer/ui/tab.ts
@@ -6,7 +6,8 @@ import { LANGUAGE_CHOICES, resolveLocale, setLocale, t, type LanguagePreference 
 import { log } from "../log";
 import type { SettingsStore } from "../settings";
 import type { PointStatus } from "../status";
-import { buildSection, button, dot, el, numberInput, toggleSwitch } from "./dom";
+import { buildSection, button, el, numberInput, toggleSwitch } from "../../ui/dom";
+import { STATUS_ICONS } from "../map-layer";
 import { getStreetNameVerdict } from "../../street-check-bridge";
 import {
   canBulkImport,
@@ -22,6 +23,21 @@ import { injectStyles } from "./styles";
 const MIN_ZOOM_BOUNDS = { min: 15, max: 22 };
 /** Statuses worth counting in the tab; NEUTRAL is the absence of a verdict, not a count. */
 const COUNTED: PointStatus[] = ["MISSING", "PRESENT", "OTHER_STREET"];
+
+/**
+ * The status marker standing next to a label; never the sole carrier of a meaning, hence
+ * the empty alt. It is the exact image the map draws, so a pill can never drift from the
+ * point it stands for.
+ *
+ * Lives here rather than in the shared builders because it reads this feature's own
+ * STATUS_ICONS, and rather than in ui/format.ts because that module holds no DOM.
+ */
+function dot(status: PointStatus): HTMLElement {
+  const img = el("img", "hn-dot");
+  img.src = STATUS_ICONS[status];
+  img.alt = "";
+  return img;
+}
 
 /**
  * The dedicated sidebar tab: master switch, state, counts, bulk import, settings, legend.
@@ -92,7 +108,7 @@ export class TabUI {
 
     this.banner.replaceChildren(this.bannerText);
     const master = el("div", "hn-master");
-    const enabledToggle = toggleSwitch(t("enable"), settings.enabled, (checked) =>
+    const enabledToggle = toggleSwitch("hn", t("enable"), settings.enabled, (checked) =>
       this.onEnabledChange(checked),
     );
     this.enabledInput = enabledToggle.querySelector("input");
@@ -205,8 +221,9 @@ export class TabUI {
       button(
         t("btnRefreshExisting"),
         () => void this.controller.refresh({ refetchExisting: true }),
+        "hn-btn",
       ),
-      button(t("btnClearCache"), () => void this.controller.reload()),
+      button(t("btnClearCache"), () => void this.controller.reload(), "hn-btn"),
     );
     return row;
   }
@@ -246,17 +263,18 @@ export class TabUI {
     });
     languageRow.append(el("span", "", t("settingsLanguage")), select);
 
-    return buildSection("⚙️", t("settingsTitle"), [
+    return buildSection("hn", "settings", t("settingsTitle"), [
       zoomRow,
-      toggleSwitch(t("settingsShowLabels"), settings.showMapLabels, (showMapLabels) => {
+      toggleSwitch("hn", t("settingsShowLabels"), settings.showMapLabels, (showMapLabels) => {
         this.settings.update({ showMapLabels });
         void this.controller.refresh();
       }),
-      toggleSwitch(t("settingsStrictMatch"), settings.strictMatching, (strictMatching) => {
+      toggleSwitch("hn", t("settingsStrictMatch"), settings.strictMatching, (strictMatching) => {
         this.settings.update({ strictMatching });
         void this.controller.refresh();
       }),
       toggleSwitch(
+        "hn",
         t("settingsExistingOnly"),
         settings.existingBuildingsOnly,
         (existingBuildingsOnly) => {
@@ -266,6 +284,7 @@ export class TabUI {
         },
       ),
       toggleSwitch(
+        "hn",
         t("settingsConfirmSingle"),
         settings.confirmSingleImport,
         (confirmSingleImport) => this.settings.update({ confirmSingleImport }),
@@ -282,7 +301,7 @@ export class TabUI {
       legend.appendChild(row);
     }
     // Open by default: it is four lines, and it is what makes the map readable at a glance.
-    return buildSection("🎨", t("legendTitle"), [legend], true);
+    return buildSection("hn", "layers", t("legendTitle"), [legend], true);
   }
 
   /**
